+++ app/controllers/api/v1/guilds_controller.rb
@@ -3,7 +3,7 @@ module Api
     class GuildsController < Api::V1::BaseController
       before_action :find_guilds, only: %i[index]
       before_action :find_guild_by_slug, only: %i[characters kick_character leave_character]
-      before_action :find_guild, only: %i[show update characters_for_request]
+      before_action :find_guild, only: %i[show update characters_for_request import_bank]
       before_action :find_guild_character, only: %i[kick_character]
       before_action :find_user_character_in_guild, only: %i[leave_character]
       before_action :find_guild_characters, only: %i[characters]
@@ -108,6 +108,21 @@ module Api
         render json: {
           characters: ActiveModelSerializers::SerializableResource.new(@characters_for_request, each_serializer: CharacterIndexSerializer).as_json[:characters]
         }, status: 200
+      end
+
+      api :POST, '/v1/guilds/:id/import_bank.json', 'Import bank'
+      param :id, String, required: true
+      param :data, String, required: true
+      error code: 401, desc: 'Unauthorized'
+      error code: 400, desc: 'Object is not found'
+      def import_bank
+        authorize! @guild, to: :bank?
+        result = ImportGuildBankData.call(guild: @guild, bank_data: params[:bank_data])
+        if result.success?
+          render json: { result: 'Bank data is importing' }, status: 200
+        else
+          render json: { result: 'Invalid bank data' }, status: 409
+        end
       end
 
       private