+++ app/interactors/update_bank_cells.rb
@@ -1,0 +1,9 @@
+# Update bank cells
+class UpdateBankCells
+  include Interactor
+
+  # required context
+  # context.bank
+  # context.cells_info
+  def call; end
+end