+++ config/routes.rb
@@ -11,6 +11,7 @@ Rails.application.routes.draw do
         get :search, on: :collection
         get :form_values, on: :collection
         get :characters_for_request, on: :member
+        post :import_bank, on: :member
       end
       resources :roles, only: %i[index]
       resources :dungeons, only: %i[index]